requirements.txt
streamlit
openai
PyPDF2
python-pptx
python-dotenv
